requirements.txt
requests
python-telegram-bot
psycopg2
configparser
openai
numpy
sqlalchemy
alembic
langdetect
pytz
pgvector
scikit-learn
python-dotenv[cli]
pytest
pytest-asyncio
python-telegram-bot[rate-limiter]